requirements.txt
click==8.0.3
Flask==2.0.2
hbase-rest-py==1.0.0
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
requests==2.5.4.1
requests-toolbelt==0.4.0
Werkzeug==2.0.2
